+++ crates/backend-uzu/src/encodable_block/mixer/delta_net/tree_verify.rs
@@ -6,10 +6,10 @@ use crate::{
 #[derive(Clone, Copy)]
 pub struct TreeVerifyNewArguments {
     pub data_type: DataType,
-    pub num_k_heads: usize,
-    pub num_v_heads: usize,
-    pub head_k_dim: usize,
-    pub head_v_dim: usize,
+    pub num_k_heads: u32,
+    pub num_v_heads: u32,
+    pub head_k_dim: u32,
+    pub head_v_dim: u32,
 }
 
 #[derive(Clone, Copy)]
@@ -21,7 +21,7 @@ pub struct TreeVerifyEncodeArguments<'a, B: Backend> {
     pub log_decay: &'a Allocation<B>,
     pub beta: &'a Allocation<B>,
     pub h0: &'a Allocation<B>,
-    pub tree_size: usize,
+    pub tree_size: u32,
 }
 
 #[cfg(test)]
